sdm845-common: DeviceExtras: Add Adreno GPU thermal throttle toggle
Yes you have to write a 0 to the node early on or else the perf HAL
will try to force it on and DeviceExtras Startup class won't restore
your preference on boot correctly.

It still might need to get manually toggled, but oh well.
This is as far as I'm willing to fight with QTI perf.